// app/types.ts
export interface IndustryCategory {
  Category: string;
  CategoryKor: string;
  SubCategory: Industry[];
}

export interface Industry {
  Industry: string;
  "Industry Ko": string;
}